Added readme and additional documentation.

+++ util/local_ee_image.py
@@ -35,8 +35,6 @@ import matplotlib.pyplot as plt
 import PyQt4
 from PyQt4 import QtCore, QtGui
 
-# The default URL to fetch tiles from.  We could pull this from the EE library,
-# however this doesn't have any other dependencies on that yet, so let's not.
 BASE_URL = 'https://earthengine.googleapis.com'
 
 TEMP_FILE_DIR = tempfile.gettempdir()
@@ -46,6 +44,17 @@ class LocalEEImage(object):
 
     # download an image locally. Caches images with the same bbox, scale and image_name
     def __init__(self, eeobject, bbox, scale, bands, image_name=None):
+        """Constructs a LocalEEImage.
+
+        Arguments:
+        eeobject -- The ee.Image object to download.
+        bbox -- A four-tuple describing the bounds to download, (lon0, lat0, lon1, lat1).
+        scale -- The scale in meters of the image to download.
+        bands -- A list containing the names of the bands in eeobject to download.
+        image_name -- A unique name for this image. If an image of the same name, bbox and
+        scale was downloaded previously, a cached image may be used. If a name is not specified,
+        no caching is performed.
+        """
         image_name = image_name.replace(' ', '') # can't have space in filename
         if image_name == None:
             image_name = uuid.uuid4()
@@ -53,7 +62,7 @@ class LocalEEImage(object):
                                        'region': apply(ee.Geometry.Rectangle, bbox).toGeoJSONString()})
         tempFile = '%s_%g,%g_%g,%g_%g.zip' % (image_name, bbox[0], bbox[1], bbox[2], bbox[3], scale)
         filename = os.path.join(TEMP_FILE_DIR, tempFile) 
-        if not os.path.isfile(filename):
+        if image_name != None and not os.path.isfile(filename):
             print 'Downloading image...'
             data = urllib2.urlopen(url)
             with open(filename, 'wb') as fp:
@@ -63,7 +72,9 @@ class LocalEEImage(object):
                     fp.write(chunk)
             print 'Download complete!'
 
+        # extract the zip file
         z = zipfile.ZipFile(filename, 'r')
+        # read the transform from the tfw file
         transform_file = z.open(image_name + '.' + bands[0] + '.tfw', 'r')
         self.transform = [float(line) for line in transform_file]
 
@@ -79,26 +90,26 @@ class LocalEEImage(object):
         self.scale      = scale
 
     def image_to_global(self, r, c):
-        '''Convert pixel coordinate to latitude and longitude'''
+        '''Convert pixel coordinate to latitude and longitude.'''
         lng = self.transform[0] * c + self.transform[4]
         lat = self.transform[3] * r + self.transform[5]
         return (lng, lat)
 
     def global_to_image(self, lon, lat):
-        '''Convert a latitude and longitude to a pixel coordinate'''
+        '''Convert a latitude and longitude to a pixel coordinate.'''
         c = (lon - self.transform[4]) / self.transform[0]
         r = (lat - self.transform[5]) / self.transform[3]
         return (r, c)
 
     def get(self, band, r, c):
-        '''Fetch the selected pixel'''
+        '''Fetch the selected pixel.'''
         return self.images[band][r, c]
 
     def get_image(self, band):
-        '''Fetch the selected band'''
+        '''Fetch the selected band as a PIL image.'''
         return self.images[band]
 
     def size(self):
-        '''Get the width and height of the image'''
+        '''Get the width and height of the image.'''
         return self.images[self.bands[0]].shape
